Add files via upload
Here I have create a Weather checking Application . This project is created using Python ,Html , Css and Javascripts. With the help of css , I have designed a beautiful background , add a background sound , with the help of Javascript and Html.
I have learned to so many things while making this project.
This project show my performance on Programming languages and web designing.